# testra/docs/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Testra API
  description: |
    Testra — One Platform. Every Test.
    RESTful JSON API for the Testra Quality Engineering Platform.
    All responses use a consistent envelope: `{ "data": ..., "meta": ..., "error": ... }`.
  version: 0.5.0
  contact:
    name: Testra Engineering
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
  - url: https://staging.testra.io/api/v1
    description: Staging
  - url: https://api.testra.io/api/v1
    description: Production
...
paths:
  /auth/register:
    post:
      tags: [Auth]
      operationId: register
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, name]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 12, description: Minimum 12 characters }
                name: { type: string }
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/AuthResult" }
        "409": { $ref: "#/components/responses/Conflict" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: Authenticate with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/AuthResult" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/refresh:
    post:
      tags: [Auth]
      operationId: refreshToken
      summary: Exchange a refresh token for a new access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        "200":
          description: New access and refresh tokens
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/AuthResult" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/me:
    get:
      tags: [Auth]
      operationId: getCurrentUser
      summary: Get the authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /organizations:
    post:
      tags: [Organizations]
      operationId: createOrganization
      summary: Create an organization
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, slug]
              properties:
                name: { type: string }
                slug: { type: string }
      responses:
        "201":
          description: Organization created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/Organization" }
        "409": { $ref: "#/components/responses/Conflict" }
        "403": { $ref: "#/components/responses/Forbidden" }
    get:
      tags: [Organizations]
      operationId: listOrganizations
      summary: List organizations for the authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Organizations
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/Organization" }
  /organizations/{id}:
    get:
      tags: [Organizations]
      operationId: getOrganization
      summary: Get an organization by ID
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Organization
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/Organization" }
        "404": { $ref: "#/components/responses/NotFound" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /workspaces:
    post:
      tags: [Workspaces]
      operationId: createWorkspace
      summary: Create a workspace
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [organization_id, name, slug]
              properties:
                organization_id: { type: string, format: uuid }
                name: { type: string }
                slug: { type: string }
      responses:
        "201":
          description: Workspace created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/Workspace" }
        "409": { $ref: "#/components/responses/Conflict" }
        "403": { $ref: "#/components/responses/Forbidden" }
    get:
      tags: [Workspaces]
      operationId: listWorkspaces
      summary: List workspaces for an organization
      security:
        - bearerAuth: []
      parameters:
        - name: organization_id
          in: query
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Workspaces
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/Workspace" }
  /workspaces/{id}:
    get:
      tags: [Workspaces]
      operationId: getWorkspace
      summary: Get a workspace by ID
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Workspace
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/Workspace" }
        "404": { $ref: "#/components/responses/NotFound" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /projects:
    post:
      tags: [Projects]
      operationId: createProject
      summary: Create a project
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [workspace_id, name, key]
              properties:
                workspace_id: { type: string, format: uuid }
                name: { type: string }
                key:
                  type: string
                  description: Short uppercase project key, e.g. "CORE"
                  pattern: "^[A-Z][A-Z0-9]{1,9}$"
                description: { type: string }
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/Project" }
        "409": { $ref: "#/components/responses/Conflict" }
        "403": { $ref: "#/components/responses/Forbidden" }
    get:
      tags: [Projects]
      operationId: listProjects
      summary: List projects for a workspace
      security:
        - bearerAuth: []
      parameters:
        - name: workspace_id
          in: query
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Projects
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/Project" }
  /projects/{id}:
    get:
      tags: [Projects]
      operationId: getProject
      summary: Get a project by ID
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/Project" }
        "404": { $ref: "#/components/responses/NotFound" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
    Conflict:
      description: Resource already exists
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
  schemas:
    Envelope:
      type: object
      properties:
        data: {}
        meta: {}
        error:
          type: ["object", "null"]
...
    AuthResult:
      type: object
      properties:
        token: { type: string, description: JWT access token (15 min expiry) }
        refresh_token: { type: string, description: Opaque refresh token for token rotation }
        user: { $ref: "#/components/schemas/User" }
    Organization:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
...